refactor: tidy up test cases in parser directory
