+++ test/helpers/style.ts
@@ -149,3 +149,24 @@ export function pElement(
     pseudoClass
   }
 }
+
+export function assertWithoutRange(result: Style, expected: Style): void {
+  expect(excludeRange(result)).toEqual(excludeRange(expected))
+}
+
+function excludeRange(obj: any): any {
+  if (Array.isArray(obj)) {
+    return obj.map(excludeRange)
+  } else if (obj === null || typeof obj !== 'object') {
+    return obj
+  }
+
+  const res: any = {}
+  Object.keys(obj).forEach(key => {
+    if (key !== 'range') {
+      const value = obj[key]
+      res[key] = excludeRange(value)
+    }
+  })
+  return res
+}
