+++ test/parser/modifier.spec.ts
@@ -1,7 +1,4 @@
-import { parseComponent } from 'vue-template-compiler'
 import { parse as parseTemplate } from 'vue-eslint-parser'
-import { parse as _parseScript } from 'babylon'
-import _parseStyle from 'postcss-safe-parser'
 import {
   modify,
   insertBefore,
@@ -9,12 +6,7 @@ import {
   remove,
   replace
 } from '@/parser/modifier'
-import { insertToTemplate } from '@/parser/template/modify'
-import { insertComponentScript } from '@/parser/script/modify'
-import { updateDeclaration } from '@/parser/style/modify'
 import { transformTemplate } from '@/parser/template/transform'
-import { Style } from '@/parser/style/types'
-import { transformStyle } from '@/parser/style/transform'
 
 describe('Modifier', () => {
   it('should insert string before specified range', () => {
@@ -107,343 +99,5 @@ describe('Modifier', () => {
     ])
     const expected = 'let ;'
     expect(actual).toBe(expected)
-  })
-})
-
-describe('Template modifier', () => {
-  it('should insert a node code', () => {
-    const code = `
-    <template>
-      <div>
-        <a href="#">Test</a>
-      </div>
-    </template>
-    `
-
-    const program = parseTemplate(code, {})
-    const ast = transformTemplate(program.templateBody!, code)
-
-    const actual = modify(code, [
-      insertToTemplate(ast, [1, 1], '<h1>Hello!</h1>')
-    ])
-    const expected = `
-    <template>
-      <div>
-        <h1>Hello!</h1>
-        <a href="#">Test</a>
-      </div>
-    </template>
-    `
-    expect(actual).toBe(expected)
-  })
-
-  it('should insert to the last children', () => {
-    const code = `
-    <template>
-      <div>
-        <a href="#">Test</a>
-      </div>
-    </template>
-    `
-
-    const program = parseTemplate(code, {})
-    const ast = transformTemplate(program.templateBody!, code)
-
-    const actual = modify(code, [
-      insertToTemplate(ast, [1, 3], '<p>Message</p>')
-    ])
-    const expected = `
-    <template>
-      <div>
-        <a href="#">Test</a>
-        <p>Message</p>
-      </div>
-    </template>
-    `
-    expect(actual).toBe(expected)
-  })
-
-  it('should insert into an empty element', () => {
-    const code = `
-    <template>
-      <div></div>
-    </template>
-    `
-
-    const program = parseTemplate(code, {})
-    const ast = transformTemplate(program.templateBody!, code)
-
-    const actual = modify(code, [
-      insertToTemplate(ast, [1, 0], '<p>Message</p>')
-    ])
-    const expected = `
-    <template>
-      <div>
-        <p>Message</p>
-      </div>
-    </template>
-    `
-    expect(actual).toBe(expected)
-  })
-
-  it('should insert into one line code', () => {
-    const code = `
-    <template>
-      <div>
-        <h1>Test</h1>
-      </div>
-    </template>
-    `
-
-    const program = parseTemplate(code, {})
-    const ast = transformTemplate(program.templateBody!, code)
-
-    const actual = modify(code, [
-      insertToTemplate(ast, [1, 1, 1], '<span>Message</span>')
-    ])
-    const expected = `
-    <template>
-      <div>
-        <h1>Test
-          <span>Message</span>
-        </h1>
-      </div>
-    </template>
-    `
-    expect(actual).toBe(expected)
-  })
-})
-
-describe('Script modifier', () => {
-  function parseScript(code: string) {
-    return _parseScript(code, {
-      sourceType: 'module'
-    }).program
-  }
-
-  it('should add child component', () => {
-    const code = `
-    import Foo from './Foo.vue'
-
-    export default {
-      components: {
-        Foo
-      }
-    }
-    `
-
-    const program = parseScript(code)
-    const actual = modify(code, [
-      insertComponentScript(program, code, 'Bar', './Bar.vue')
-    ])
-
-    const expected = `
-    import Foo from './Foo.vue'
-    import Bar from './Bar.vue'
-
-    export default {
-      components: {
-        Foo,
-        Bar
-      }
-    }
-    `
-    expect(actual).toBe(expected)
-  })
-
-  it('should handle the existing trailing comma', () => {
-    const code = `
-    import Foo from './Foo.vue'
-
-    export default {
-      components: {
-        Foo,
-      }
-    }
-    `
-
-    const program = parseScript(code)
-    const actual = modify(code, [
-      insertComponentScript(program, code, 'Bar', './Bar.vue')
-    ])
-
-    const expected = `
-    import Foo from './Foo.vue'
-    import Bar from './Bar.vue'
-
-    export default {
-      components: {
-        Foo,
-        Bar,
-      }
-    }
-    `
-    expect(actual).toBe(expected)
-  })
-
-  it('should add into empty component options', () => {
-    const code = `
-    export default {
-      components: {
-      }
-    }
-    `
-
-    const program = parseScript(code)
-    const actual = modify(code, [
-      insertComponentScript(program, code, 'Foo', './Foo.vue')
-    ])
-
-    const expected = `
-    import Foo from './Foo.vue'
-    export default {
-      components: {
-        Foo
-      }
-    }
-    `
-    expect(actual).toBe(expected)
-  })
-
-  it('should add component options if not exist', () => {
-    const code = `
-    export default {
-    }
-    `
-
-    const program = parseScript(code)
-    const actual = modify(code, [
-      insertComponentScript(program, code, 'Foo', './Foo.vue')
-    ])
-
-    const expected = `
-    import Foo from './Foo.vue'
-    export default {
-      components: {
-        Foo
-      }
-
-    }
-    `
-    expect(actual).toBe(expected)
-  })
-
-  it('should only insert into script block', () => {
-    const code = `
-<template>
-  <p>Hi</p>
-</template>
-
-<script>
-export default {
-}
-</script>
-`
-
-    const { script } = parseComponent(code, { pad: 'space' })
-    const scriptCode = script!.content
-    const program = parseScript(scriptCode)
-    const actual = modify(scriptCode, [
-      insertComponentScript(program, scriptCode, 'Foo', './Foo.vue')
-    ])
-
-    const spacify = (matched: string) => matched.replace(/./g, ' ')
-
-    const expected = `
-<template>
-  <p>Hi</p>
-</template>
-
-<script>
-import Foo from './Foo.vue'
-export default {
-  components: {
-    Foo
-  }
-
-}
-`.replace(/<template>[\S\s]+<script>/, spacify)
-
-    expect(actual).toBe(expected)
-  })
-})
-
-describe('Style modifier', () => {
-  function parseStyle(code: string): Style[] {
-    return [transformStyle(_parseStyle(code), code, 0)]
-  }
-
-  it('should modify specified declaration', () => {
-    const code = `
-    p {
-      color: red;
-      font-size: 22px;
-    }
-    `
-
-    const styles = parseStyle(code)
-    const res = modify(code, [
-      updateDeclaration(styles, {
-        path: [0, 0, 1],
-        prop: 'font-weight',
-        value: 'bold',
-        important: true
-      })
-    ])
-    const expected = `
-    p {
-      color: red;
-      font-weight: bold !important;
-    }
-    `
-    expect(res).toBe(expected)
-  })
-
-  it('should partially update declaration', () => {
-    const code = `
-    p {
-      color: red;
-      font-size: 22px;
-    }
-    `
-
-    const styles = parseStyle(code)
-    const res = modify(code, [
-      updateDeclaration(styles, {
-        path: [0, 0, 1],
-        value: '24px'
-      })
-    ])
-    const expected = `
-    p {
-      color: red;
-      font-size: 24px;
-    }
-    `
-    expect(res).toBe(expected)
-  })
-
-  it('should do nothing when the declaration is not found', () => {
-    const code = `
-    p {
-      color: red;
-      font-size: 22px;
-    }
-    `
-
-    const styles = parseStyle(code)
-    const res = modify(code, [
-      updateDeclaration(styles, {
-        path: [0, 0, 2],
-        value: '24px'
-      })
-    ])
-    const expected = `
-    p {
-      color: red;
-      font-size: 22px;
-    }
-    `
-    expect(res).toBe(expected)
   })
 })
