+++ test/parser/script/child-component.spec.ts
@@ -1,184 +1,7 @@
 import * as path from 'path'
 import { parse } from 'babylon'
-import { Prop, Data, ChildComponent } from '@/parser/script/types'
-import {
-  extractProps,
-  extractData,
-  extractChildComponents
-} from '@/parser/script/manipulate'
-
-describe('Script props parser', () => {
-  it('should extract props', () => {
-    const code = `
-    export default {
-      props: {
-        foo: String,
-        bar: {
-          type: Number,
-          default: 42
-        }
-      }
-    }`
-
-    const { program } = parse(code, { sourceType: 'module' })
-    const extracted = extractProps(program)
-    const expected: Prop[] = [
-      {
-        name: 'foo',
-        type: 'String',
-        default: undefined
-      },
-      {
-        name: 'bar',
-        type: 'Number',
-        default: 42
-      }
-    ]
-    expect(extracted).toEqual(expected)
-  })
-
-  it('should extract props from array definition', () => {
-    const code = `export default { props: ['test'] }`
-
-    const { program } = parse(code, { sourceType: 'module' })
-    const extracted = extractProps(program)
-    const expected: Prop[] = [
-      {
-        name: 'test',
-        type: 'any',
-        default: undefined
-      }
-    ]
-    expect(extracted).toEqual(expected)
-  })
-
-  it('should extract props from Vue.extend defintion', () => {
-    const code = `export default Vue.extend({ props: ['foo'] })`
-
-    const { program } = parse(code, { sourceType: 'module' })
-    const extracted = extractProps(program)
-    const expected: Prop[] = [
-      {
-        name: 'foo',
-        type: 'any',
-        default: undefined
-      }
-    ]
-    expect(extracted).toEqual(expected)
-  })
-
-  it('should extract default props from function', () => {
-    const code = `
-    export default {
-      props: {
-        foo: {
-          type: String,
-          default: () => 'test'
-        }
-      }
-    }`
-
-    const { program } = parse(code, { sourceType: 'module' })
-    const extracted = extractProps(program)
-    const expected: Prop[] = [
-      {
-        name: 'foo',
-        type: 'String',
-        default: 'test'
-      }
-    ]
-    expect(extracted).toEqual(expected)
-  })
-})
-
-describe('Script data parser', () => {
-  it('should extract data', () => {
-    const code = `
-    export default {
-      data () {
-        return { foo: 'test' }
-      }
-    }`
-
-    const { program } = parse(code, { sourceType: 'module' })
-    const extracted = extractData(program)
-    const expected: Data[] = [
-      {
-        name: 'foo',
-        default: 'test'
-      }
-    ]
-    expect(extracted).toEqual(expected)
-  })
-
-  it('should extract data from arrow function', () => {
-    const code = `
-    export default {
-      data: () => ({
-        abc: 123,
-        cde: true
-      })
-    }`
-
-    const { program } = parse(code, { sourceType: 'module' })
-    const extracted = extractData(program)
-    const expected: Data[] = [
-      {
-        name: 'abc',
-        default: 123
-      },
-      {
-        name: 'cde',
-        default: true
-      }
-    ]
-    expect(extracted).toEqual(expected)
-  })
-
-  it('should extract array data', () => {
-    const code = `
-    export default {
-      data: () => ({
-        foo: ['test', 123]
-      })
-    }`
-
-    const { program } = parse(code, { sourceType: 'module' })
-    const extracted = extractData(program)
-    const expected: Data[] = [
-      {
-        name: 'foo',
-        default: ['test', 123]
-      }
-    ]
-    expect(extracted).toEqual(expected)
-  })
-
-  it('should extract object data', () => {
-    const code = `
-    export default {
-      data: () => ({
-        obj: {
-          foo: 'test',
-          bar: 123
-        }
-      })
-    }`
-
-    const { program } = parse(code, { sourceType: 'module' })
-    const extracted = extractData(program)
-    const expected: Data[] = [
-      {
-        name: 'obj',
-        default: {
-          foo: 'test',
-          bar: 123
-        }
-      }
-    ]
-    expect(extracted).toEqual(expected)
-  })
-})
+import { ChildComponent } from '@/parser/script/types'
+import { extractChildComponents } from '@/parser/script/manipulate'
 
 describe('Script components parser', () => {
   const hostPath = '/path/to/Component.vue'
